Incorporate review feedback
