+++ tests/tensor/rewriting/test_ofg.py
@@ -13,10 +13,10 @@ from pytensor.compile.builders import OpFromGraph
 def test_alloc_diag_inlined():
     x = pt.tensor("x", shape=(None,))
 
-    # pt.diag calls AllocDiag, which inherits from OpFromGrpah
     z = pt.diag(x)
-    f = pytensor.function([x], z)
+    assert isinstance(z.owner.op, OpFromGraph)
 
+    f = pytensor.function([x], z)
     nodes = f.maker.fgraph.apply_nodes
 
     assert not any(isinstance(node.op, OpFromGraph) for node in nodes)
